feat: Add UserController with GET endpoints
- Create UserController with GET /api/user/{id} and GET /api/user/me
- Fix Hibernate physical naming strategy to use table names as-is (USERS, RENTALS, MESSAGES)
- Fix ID types: use Integer (matching MySQL int type) throughout the application

+++ backend/src/main/java/com/chatop/api/service/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
     /**
      * Get user by ID
      */
-    public UserDto getUserById(Long id) {
+    public UserDto getUserById(Integer id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
